Move all files to src folder so that access can only be granted via exports
