+++ gpt4all-bindings/dart/lib/gpt4all.dart
@@ -1,2 +1,2 @@
-export 'package:gpt4all/llmodel.dart';
-export 'package:gpt4all/llmodel_generation_config.dart';
+export 'package:gpt4all/src/llmodel.dart';
+export 'package:gpt4all/src/llmodel_generation_config.dart';
