+++ gpt4all-bindings/dart/lib/src/llmodel.dart
@@ -1,7 +1,7 @@
 import 'dart:ffi' as ffi;
 import 'dart:io';
 import 'package:ffi/ffi.dart';
-import 'package:gpt4all/llmodel_generation_config.dart';
+import 'package:gpt4all/src/llmodel_generation_config.dart';
 import 'package:gpt4all/src/llmodel_error.dart';
 import 'package:gpt4all/src/llmodel_library.dart';
 import 'package:gpt4all/src/llmodel_prompt_context.dart';
